aoc/2021/src/test/java/aoc/aoc2021/Day2DiveTest.java
package aoc.aoc2021;

import aoc.common.AocFiles;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.CsvSource;

import java.io.IOException;

import static org.assertj.core.api.Assertions.assertThat;

@DisplayName("2021: Day 2: Dive!")
class Day2DiveTest {

    @ParameterizedTest
    @CsvSource({"150, day2-demo1.txt",
            "1728414, day2.txt"})
    void problem1(int expected, String fileName) throws IOException {
        var inputLines = AocFiles.readAllLines(fileName);
        assertThat(new Day2Dive(inputLines).problem1()).isEqualTo(expected);
    }

    @ParameterizedTest
    @CsvSource({"900, day2-demo1.txt",
            "1765720035, day2.txt"})
    void problem2(int expected, String fileName) throws IOException {
        var inputLines = AocFiles.readAllLines(fileName);
        assertThat(new Day2Dive(inputLines).problem2()).isEqualTo(expected);
    }
}